requirements.txt
langchain
langchain_openai
streamlit
langchain_community
pypdf
langchain-text-splitters
chromadb
faiss-cpu
arxiv
wikipedia
python-dotenv

